// src/features/recruiter/annonces/common/enums/contract.enum.ts
export enum ContractScheduleType {
  FIXED = "fixed",
  RANGE = "range",
  MAXIMUM = "maximum",
  MINIMUM = "minimum",
}

export enum ContractDurationUnit {
  DAYS = "days",
  WEEKS = "weeks",
  MONTHS = "months",
}

export enum ContractType {
  FULL_TIME = "full-time",
  PART_TIME = "part-time",
  CDI = "cdi",
  INTERIM = "interim",
  CDD = "cdd",
  FREELANCE = "freelance",
  INTERNSHIP = "internship",
  APPRENTICESHIP = "apprenticeship",
}